ai-scraper/requirements.txt
langchain-huggingface
scrapegraphai
gradio
playwright
python-dotenv
